% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt.R, R/elastic.R, R/lasso.R, R/limma.R,
%   R/package.R, R/regression.R, R/ridge.R
\name{cltApp}
\alias{cltApp}
\alias{elasticApp}
\alias{lassoApp}
\alias{limmaApp}
\alias{Apps}
\alias{regressionApp}
\alias{ridgeApp}
\title{Shiny apps to demonstrate statistical principles.}
\usage{
cltApp()

elasticApp()

lassoApp()

limmaApp()

regressionApp()

ridgeApp()
}
\description{
Shiny apps to demonstrate statistical principles.
}
\section{Functions}{
\itemize{
\item \code{cltApp}: A demonstration of the central limit theorem.

\item \code{elasticApp}: A demonstration of elastic net regression.

\item \code{lassoApp}: A demonstration of LASSO regression.

\item \code{limmaApp}: A demonstration of moderated t-tests.

\item \code{regressionApp}: A demonstration of linear regression principles.

\item \code{ridgeApp}: A demonstration of ridge regression.
}}

\examples{
\dontrun{
 regressionApp()
 limmaApp()
 ridgeApp()
 lassoApp()
 elasticApp()
}
}
