% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_effect_size.R
\name{climate_effect_app}
\alias{climate_effect_app}
\title{Visualising effect sizes through the weather}
\usage{
climate_effect_app()
}
\description{
A shiny app to help visualise effect sizes.
The app lets you choose which months to compare in a city, and shows a boxplot
of the mean monthly temperatures, the effect size (Cohen's D),
and the required sample size to detect an effect this large with
the required power at the requested significance level.
}
\examples{
\dontrun{
climate_effect_app()
}

}
