% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{colorSchemes}
\alias{colorSchemes}
\title{Color Schemes}
\format{An object of class \code{"data.frame"}; see \code{\link[shinyStats]{densityChartWMean}}.}
\source{
\href{http://www.jordandeherrera.com}{Verasite Template}
}
\usage{
data(colorSchemes)
}
\description{
Color schemes for use with graphs
}
\examples{
data(colorSchemes)
\donttest{iplotCurves(phe, times)}
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{http://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
