% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compareDensityChartWMean}
\alias{compareDensityChartWMean}
\title{Comparison Density Chart with Mean and Median}
\usage{
compareDensityChartWMean(yValue1, yValue2, yValue1Name, yValue2Name, title,
  subTitle, xAxisTitle, colorType)
}
\arguments{
\item{yValue1}{The quantitative field for series 1 from which to create the density chart.}

\item{yValue1Name}{The label for the distribution of the density chart for the first series.}

\item{yValue2Name}{The label for the distribution of the density chart for the second series.}

\item{title}{The title of the chart}

\item{subTitle}{The subtitle of the chart}

\item{xAxisTitle}{The x Axis title}

\item{colorType}{Specify the theme to use for the chart (either Verasite or Uber)}

\item{yValue1}{The quantitative field for series 2 from which to create the density chart.}
}
\value{
Returns a highchart object
}
\description{
This is a function that creates a density chart based on raw values in a population.
}
\examples{
\dontrun{
densityChartWMean(mtcars$cyl,"Density of Cylinders in MT Cars Dataset","Example of density chart","Cylinders")
}
}
