% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_data.R
\name{count_app}
\alias{count_app}
\title{Count app}
\usage{
count_app()
}
\description{
A shiny app to explore distributions for count data,
with options for the Poisson and Negative binomial distributions,
with or without zero inflation.
}
\examples{
\dontrun{
count_app()
}

}
