% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{densityChartWMean}
\alias{densityChartWMean}
\title{Density Chart with Meank}
\usage{
densityChartWMean(yValue, title, subTitle, xAxisTitle, colorType)
}
\arguments{
\item{yValue}{The quantitative field from which to create the density chart.}

\item{title}{The title of the chart}

\item{subTitle}{The subtitle of the chart}

\item{xAxisTitle}{The x Axis title}

\item{colorType}{Specify the theme to use for the chart (either Verasite or Uber)}
}
\value{
Returns a highchart object
}
\description{
This is a function that creates a density chart based on raw values in a population.
}
\examples{
\dontrun{
densityChartWMean(mtcars$cyl,"Density of Cylinders in MT Cars Dataset","Example of density chart","Cylinders")
}
}
