% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f-dist.R
\name{f_dist_app}
\alias{f_dist_app}
\title{F distribution}
\usage{
f_dist_app()
}
\description{
A shiny app to visualise how f-distribution changes with the number of
numerator and denominator degrees of freedom.
}
\examples{
\dontrun{
f_dist_app()
}
}
