% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric Mean}
\source{
\href{https://towardsdatascience.com/on-average-youre-using-the-wrong-average-geometric-harmonic-means-in-data-analysis-2a703e21ea0}{592862722639. “On Average, You're Using the Wrong Average: Geometric & Harmonic Means in Data Analysis.” Towards Data Science, Towards Data Science, 28 Jan. 2018, towardsdatascience.com/on-average-youre-using-the-wrong-average-geometric-harmonic-means-in-data-analysis-2a703e21ea0.}
}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{The quantitative field from which to calculate the geometric mean.}
}
\value{
Returns a numeric which is the geometric mean
}
\description{
This is a function taken from the blog \href{https://stackoverflow.com/questions/2602583/geometric-mean-is-there-a-built-in}{Stack Overflow} that calculates the geometric mean of a set of numbers.  Along with the \code{\link[shinyStats]{h_mean}} and \code{\link[base]{mean}} functions, they provide the three most basic functions of "average." Typically, the arithmetic mean should be used for linear relationships, the geometric mean should be used for exponential or multiplicative relationships, and the harmonic mean should be used for reciprocal relationships.
}
\examples{
\dontrun{gm_mean(mtcars$mpg)}
}
