% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{h_mean}
\alias{h_mean}
\title{Harmonic Mean}
\source{
\href{https://towardsdatascience.com/on-average-youre-using-the-wrong-average-geometric-harmonic-means-in-data-analysis-2a703e21ea0}{592862722639. “On Average, You're Using the Wrong Average: Geometric & Harmonic Means in Data Analysis.” Towards Data Science, Towards Data Science, 28 Jan. 2018, towardsdatascience.com/on-average-youre-using-the-wrong-average-geometric-harmonic-means-in-data-analysis-2a703e21ea0.}
}
\usage{
h_mean(x)
}
\arguments{
\item{x}{The quantitative field from which to calculate the harmonic mean.}
}
\value{
Returns a numeric which is the harmonic mean
}
\description{
This is a function taken from the blog \href{https://statistic-on-air.blogspot.com/2009/07/geometric-and-harmonic-means-in-r.html}{Statistics on aiR} that calculates the harmonic mean of a set of numbers.  Along with the \code{\link[shinyStats]{gm_mean}} and \code{\link[base]{mean}} it provides the three most basic functions of "average." Typically, the arithmetic mean should be used for linear relationships, the geometric mean should be used for exponential or multiplicative relationships, and the harmonic mean should be used for reciprocal relationships.
}
\examples{
\dontrun{h_mean(mtcars$mpg)}
}
