% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{methylation}
\alias{methylation}
\title{Human methylation data on immunomagnetic sorted peripheral adult blood cells}
\format{
A \linkS4class{GenomicRatioSet} with 10,000 features and 37 samples
along with associated metadata.
}
\source{
\url{https://bioconductor.org/packages/3.12/data/experiment/html/FlowSorted.Blood.EPIC.html}
}
\usage{
methylation
}
\description{
A random subset of 10,000 features from the Bioconductor package 
"FlowSorted.Blood.EPIC" from Bioconductor 3.13.
}
\details{
The FlowSorted.Blood.EPIC object is based in samples assayed by Brock 
Christensen and colleagues; for details see Salas et al. 2018.
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE110554}.
}
\keyword{datasets}
