% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misreported_cohens_d_app.R
\name{misreported_cohens_d_app}
\alias{misreported_cohens_d_app}
\title{Misreported Cohen's D app}
\usage{
misreported_cohens_d_app()
}
\description{
A shiny app that helps visualise what happens when Cohen's D is misreported.
}
\examples{
\dontrun{
misreported_cohens_d_app()
}
}
