% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_line_ss.R
\name{regression_line_ss_app}
\alias{regression_line_ss_app}
\title{Regression line ss app}
\usage{
regression_line_ss_app()
}
\description{
A shiny app to help understand residuals of a linear model.
Draw the best fit line by click twice, once at each end,
then choose how to display the residuals.
Click again to draw a new line.
}
\examples{
\dontrun{
regression_line_ss_app()
}

}
