% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_alpha}
\alias{set_alpha}
\title{Set alpha channel of colour}
\usage{
set_alpha(colour, alpha = 1)
}
\arguments{
\item{colour}{A hex colour code}

\item{alpha}{A value between 0 and 1}
}
\value{
A hex colour code with the alpha channel set
}
\description{
Set alpha channel of colour
}
\examples{
set_alpha("#832424", 0.5)
}
