#' shinyStats Package Documentation
#'
#'
#'
#'
#' statistical distributions and power laws.
#'
#' \subsection{Statistical Distributions}{
#'
#' \code{\link{densityChartWMean}} creates a density chart with vertical plot lines for percentiles and means of the distribution
#'
#' }
#'
#' \subsection{Power Laws}{
#'
#' \code{\link{percRank}} - calculate percent rank of a quantitative field
#'
#' }
#'
"_PACKAGE"

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyBoots.R
\docType{package}
\name{shinyStats-package}
\alias{shinyStats}
\alias{shinyStats-package}
\title{shinyStats Package Documentation}
\description{
This package contains functions to make it easier to create common statistical analyses for exploratory analysis.
}
\details{

The package contains functions that fall into the following general categories:  statistical distributions and power laws.

\subsection{Statistical Distributions}{

\code{\link{densityChartWMean}} creates a density chart with vertical plot lines for percentiles and means of the distribution

}

\subsection{Power Laws}{

\code{\link{percRank}} - calculate percent rank of a quantitative field

}


}
\author{
\strong{Maintainer}: Jordan DeHerrera \email{jordan.deherrera@gmail.com}

}
