% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_error_app.R
\name{standard_error_app}
\alias{standard_error_app}
\title{Standard error app}
\usage{
standard_error_app()
}
\description{
A shiny app to visualise how the population standard deviation and the number
samples affects the uncertainty of the mean.
}
\examples{
\dontrun{
standard_error_app()
}
}
