% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survivalCurve}
\alias{survivalCurve}
\title{Survival Curve}
\usage{
survivalCurve(alpha, beta, t = 96)
}
\arguments{
\item{alpha}{Alpha parameter}

\item{beta}{Beta parameter}

\item{t}{iterations}
}
\value{
Returns a vector of survival percent estimates
}
\description{
This function creates a survival curve based on an \code{\link{sBG.est}} model parameters.
}
\examples{
\dontrun{
library(dplyr)
library(titanic)
titanic::titanic_train \%>\%
mutate(Deceased = ifelse(Survived == 1,0,1)) \%>\%
group_by(Age = round(Age,0)) \%>\%
summarise(Survived = sum(Survived), Deceased = sum(Deceased)) \%>\% {
sBG.est(
.$Survived[2:nrow(.)],
.$Deceased[2:nrow(.)]
)$par} \%>\% {
survivalCurve(.[1], .[2], 50)
}
}
}
