% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dist_app.R
\name{t_dist_app}
\alias{t_dist_app}
\title{t distribution}
\usage{
t_dist_app()
}
\description{
A shiny app to visualise how t-distribution changes with the number of
degrees of freedom.
}
\examples{
\dontrun{
f_dist_app()
}
}
