\docType{package}
\name{shinyStats-package}
\alias{shinyStats}
\alias{shinyStats-package}
\title{shinyStats Package Documentation}
\description{
This package contains functions to make it easier to create common statistical analyses for exploratory analysis.
}
\details{

The package contains functions that fall into the following general categories:  statistical distributions and power laws.

\subsection{Statistical Distributions}{

\code{\link{densityChartWMean}} creates a density chart with vertical plot lines for percentiles and means of the distribution

}

\subsection{Power Laws}{

\code{\link{percRank}} - calculate percent rank of a quantitative field

\code{\link{sBG.est}} - estimates parameters of shifted beta gamma curves

\code{\link{survivalCurve}} - creates a survival curve based off of beta model

\code{\link{h_mean}} - calculate the harmonic mean for a set of numbers

\code{\link{gm_mean}} - calculate the geometric mean for a set of numbers

}


}
\author{
\strong{Maintainer}: Jordan DeHerrera \email{jordan.deherrera@gmail.com}

}
